+++ app/src/main/java/io/rafaelribeiro/spendless/core/data/TransactionCreator.kt
@@ -1,5 +1,6 @@
-package io.rafaelribeiro.spendless.data.repository
+package io.rafaelribeiro.spendless.core.data
 
+import io.rafaelribeiro.spendless.data.repository.UserPreferences
 import io.rafaelribeiro.spendless.domain.CurrencySymbol
 import io.rafaelribeiro.spendless.domain.DecimalSeparator
 import io.rafaelribeiro.spendless.domain.ExpenseFormat
@@ -8,66 +9,16 @@ import io.rafaelribeiro.spendless.domain.ThousandSeparator
 import io.rafaelribeiro.spendless.domain.Transaction
 import io.rafaelribeiro.spendless.domain.TransactionCategory
 import io.rafaelribeiro.spendless.domain.TransactionFormatter
-import io.rafaelribeiro.spendless.domain.TransactionRepository
 import io.rafaelribeiro.spendless.domain.TransactionType
 import io.rafaelribeiro.spendless.domain.toUIModel
 import io.rafaelribeiro.spendless.presentation.screens.dashboard.TransactionUiModel
-import kotlinx.coroutines.flow.Flow
-import kotlinx.coroutines.flow.flow
 import java.time.Instant
 import java.time.LocalDateTime
 import java.time.ZoneId.systemDefault
 import java.time.format.DateTimeFormatter
 import java.time.temporal.ChronoUnit
 import java.util.Locale
-import javax.inject.Inject
 import kotlin.random.Random
-
-/**
- * Fake implementation of [TransactionRepository] to be used in tests.
- * Change DI's [TransactionRepository] implementation to this one to use it in tests for convenience.
- */
-class FakeTransactionRepository @Inject constructor() : TransactionRepository {
-    override fun getBalance(): Flow<Double?> {
-        return flow {
-            emit(1000.toDouble())
-        }
-    }
-
-    override fun getAllTransactions(): Flow<List<Transaction>> {
-        return flow {
-            emit(TransactionCreator.createTransactions(40))
-        }
-    }
-
-    override fun getLatestTransactions(): Flow<List<Transaction>> {
-        return flow {
-            emit(TransactionCreator.createTransactions(10))
-        }
-    }
-
-    override fun getTotalAmountLastWeek(): Flow<Double?> {
-        return flow {
-            emit(10000.toDouble())
-        }
-    }
-
-    override fun getBiggestTransaction(): Flow<Transaction?> {
-        return flow {
-            emit(TransactionCreator.createTransaction())
-        }
-    }
-
-    override fun getMostPopularCategory(): Flow<TransactionCategory?> {
-        return flow {
-            emit(TransactionCategory.FOOD)
-        }
-    }
-
-    override suspend fun saveTransaction(transaction: Transaction) {}
-
-    override suspend fun deleteAllTransactions() {}
-}
 
 /**
  * Helper class to create transactions for tests.
