+++ app/src/main/java/io/rafaelribeiro/spendless/presentation/screens/dashboard/components/DashboardTransactionItem.kt
@@ -31,7 +31,7 @@ import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import io.rafaelribeiro.spendless.data.repository.TransactionCreator
+import io.rafaelribeiro.spendless.core.data.TransactionCreator
 import io.rafaelribeiro.spendless.domain.TransactionType.EXPENSE
 import io.rafaelribeiro.spendless.domain.TransactionType.INCOME
 import io.rafaelribeiro.spendless.presentation.screens.dashboard.TransactionUiModel
